Implmeent auth router according to nextauth v4
