+++ ui/.gitignore
@@ -42,3 +42,4 @@ next-env.d.ts
 /playwright/.cache/
 .env
 dotenv
+.env*
