+++ ui/src/app/page.tsx
@@ -1,16 +1,12 @@
-import {
-  Button,
-  Input,
-  Gradient,
-  Text,
-  Socials,
-  Box,
-  Stack,
-  Password,
-  Row,
-} from "../components/ui";
+import { Text } from "@uug-ai/ui";
 import React from "react";
 
-export default function App() {
-  return <div>App</div>;
+export default async function App() {
+  return (
+    <div>
+      <Text as="h1" size="5xl">
+        APP
+      </Text>
+    </div>
+  );
 }
